teropi/requirements.txt
pyftpdlib
pyzmq
